requirements.txt
# requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
plotly==5.17.0
nltk==3.8.1
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0
gradio